next video iteration
